backend/requirements.txt
-r stub/requirements.txt
DBUtils
PyMySQL
connexion<3.0.0
numpy
scikit-learn
sqlalchemy
statsmodels
cryptography
flask_cors